feat(weather): support multiple sources
Add support for open-meteo.com as a weather source. Prefer this source
because it doesn't _require_ an api-key. _If_ a weatherapi.com api-key
is present then that source will be preferred.

Modifications were made to the concept of location to support this
change. Both of these weather sources will accept a lat/long pair so
that is what we now use. We no longer allow the user to set a "city,
state" as location. Location can be inferred (if unset) or set by postal
code.

Resolves #57