refactor: towards (more) simplicity (#93)
- validity

Introduce `is_valid` for `Fetchable`. This allows each implementation
to determine internally whether or not it in a proper state to make a
request. Previously some of this logic (eg. do we have an api-key) had
bled out of the implementation.

- context

Push `config` and `location` down to weather providers. Previously the
caller of these providers had to differentiate what their needs were
(api-key? unit of measure?).

- trait object

It's unconventional to have a `const` defined in a trait, move `URL`
to - `fn url()`.

- dead code

Remove `location::HttpClient`, it was effectively replaced by
`api::Fetchable`.