feat(location): add caching
Use an embedded SQLite client to capture location details for future
reference. At this time there is very little (no?) value to this caching
mechanism because a user has to first set their region/location in order
for the cache to be read. The act of setting a region though will
persist location details in `config.toml` (effectively a cache).

Caching will come into effect once #122 is implemented.

Resolves #120